Add functionality to create a tester user
The UserServiceImpl and KeycloakService have been updated to provide an ability to create a tester user. A new function 'createTester' has been added to KeycloakService to create a tester user with Super Admin privileges. The UserServiceImpl now checks if such a tester user exists before creation, and if not, then proceeds to create one.