Remove non-null restrictions and comment out user initialization
The non-null restrictions on 'name' and 'code' fields in the 'KaspiCity' entity class have been removed, allowing these fields to have null values. Also, the code for user initialization in the 'Initializer' class has been commented out. This should prevent the automatic creation of users during the initialization process. The commit improves the flexibility of 'KaspiCity' entity and halts the automatic user generation for testing purposes.