Add storage related functionalities to SupplyService
The SupplyService interface has been enhanced to support retrieval of supplies with respect to a storage employee, based on either the employeeId or the keycloakId. Additional methods are added to retrieve supply products by keycloakId and supplyId. SelectedTime has been added to Supply entity, and a new ProductStorageResponse DTO has been introduced. Changes also include query methods in SupplyRepository and StoreEmployeeRepository, and enhanced methods in SupplyServiceImpl.