Update Supply createSupply return type and response format
The changes focus on updating the SupplyService createSupply method to return a long id of the created supply. It also modifies the createSupply endpoint in SupplyController to return a Map with the created supply id as a response. This change provides more information to the client about the created resource.