Update UserService and add OrderService with enhancements
This commit updates the UserServiceImpl to set a default seller name. It introduces new functionality in the form of an OrderService, which allows fetching of admin and seller orders based on Keycloak ID. Additionally, it updates the KaspiOrder and KaspiStore entities for better address handling and order-user assignment. This update also contains changes in application settings and improvements to schedulers and controllers.