Update order and city processing
The main classes 'OrderServiceImpl', 'OrderService', 'ValueConstants', and 'Scheduler' have been updated to improve the handling of city and order data from the Kaspi API. The new implementation ensures city data from the Kaspi API is updated every week and order information is updated every five minutes. This information is fetched and processed for each token found in the database.