Update Order service and KaspiStore retrieval
Revised the OrderServiceImpl to remove the timeout restriction and refactor the method for assigning parameters to the KaspiStore. Edited the types of the latitude and longitude fields in the KaspiStore and updated the associated lookup query. Also, certain application settings are updated, including the change of the database domain to localhost and the disabling of user and city syncing.