Add trade price to products and date filters for orders retrieval
This update introduces a new trade price property for products. It also changes the order retrieval methods, 'getSellerOrdersByKeycloakId' and 'getAdminOrdersByKeycloakId', to accept two new parameters: a start date and an end date. This means that the system will now be able to retrieve orders within a specified date range. The processing of product prices in different cities has also been adjusted accordingly.