Refactor OrderServiceImpl and SupplyBoxProductsRepository
Changed the search logic in OrderServiceImpl and modified the database query in SupplyBoxProductsRepository. Removed specific time check in the processing order block in OrderServiceImpl. Instead of throwing an exception when a product not found in the kaspi store, the code now simply returns. Rewrote the native SQL query in the SupplyBoxProductsRepository to a JPQL query for better readability and maintainability.