Add order details retrieval functionality
Enhanced the OrderService to add two new methods for retrieving the details of an order, specifically for admin and seller roles. These new functionalities build upon the on-demand fetch ethics, bringing forth a more detailed view regarding order-related specifics. Additionally, minor adjustments have been made to the OrderResponse class with the removal of the ID property. Lastly, a new Order