Add SupplyBoxRepository and implement getSupplyByBox
SupplyBoxRepository has been created and a new method getSupplyByBox has been added in the SupplyService and SupplyController. The changes also extend the SupplyService and its implementation to support fetching supplies based on a boxVendorCode. This allows fetching the Supply related to a specific SupplyBox using its vendor code. The role of the user is also checked and the result only returns if the user has the 'SUPER_ADMIN' role.