Update SecurityConfig and refactor repositories
The SecurityConfig class has been updated to allow requests to the "/actuator/**" path. Moreover, changes have been made to the KaspiStoreRepository, specifically, methods and imports have been added related to pagination, and a finder method was removed. Comment added in BoxTypesController querying the need for pagination in the getAll method.