Update product processing in orders and supplies
The commit introduces a StoreCellProductRepository in OrderServiceImpl for providing cell codes in order details. It also adjusts the JoinColumn names in StoreCellProduct entity. Changes are made in SupplyServiceImpl to update product state and accepted time while processing supplies by employees. In SupplyController, the ResponseEntity status is updated to "CREATED" from "OK" after processing supplies. An additional function is provided in StoreCellProductRepository to find by SupplyBoxProductId.