Refactor StoreCellServiceImpl to use UserRepository
The StoreCellServiceImpl class was updated to include the UserRepository instance and refactor the 'changeStoreCell' method. This change simplifies checking for the store of a cell against the KeycloakId, therefore improving efficiency. The check for store employee's existence and their matching to the cell has been removed, reducing unnecessary code complexity.