Added method to update product price
A new method called 'changePrice' has been added to 'ProductServiceImpl' to allow the product price to be updated. An API endpoint has also been added to utilize this functionality. The availability check for the product has been modified to consider the price as non-zero for availability. Additionally, 'flyway-core' dependency reintroduced in build.gradle.