Refactor product price change functionality for multiple cities
The product price change method has been greatly refactored. Previously, the method was applicable for change in price of a product in only one city. Now, a new request model has been introduced to allow setting different prices for a product in multiple cities simultaneously. To achieve this, the method now iterates over a list of prices and their respective cities from this model.