Refactored OrderServiceImpl and introduced KaspiOrderMapper
Redesigned the implementation of OrderServiceImpl by introducing KaspiOrderMapper to reduce the complexity of the code. Also created new mapper classes for KaspiDeliveryAddress and KaspiStore. This refactor aims to enhance code readability and maintainability. Removed redundant repository calls and streamlined data retrieval for orders and order products. Configuration for the application's mode has also been moved to an environment variable for improved versatility.