Refactor log and modify product-prices relationship
In the ProductServiceImpl.java, the logging information has been moved to a point after processing the product prices. Also, a cascade type ALL has been added to the OneToMany relationship between product and prices in the Product.java entity. This will ensure that any operation performed on a product will cascade to its associated prices.