Added feature for product search by parameters
The code updates involve implementing a feature that allows product search by multiple parameters. This includes the creation of new classes for search request and response, and a new method in the ProductService interface. Corresponding implementations were made in the ProductServiceImpl class. Various repositories are also updated to support this new feature.