Add product count information to seller stats
A new feature was implemented that provides count information of the seller's products. This was achieved by adding a new method in the StatisticsService interface and its implementation in StatisticsServiceImpl. Additionally, a new endpoint was added in the StatisticsController to expose this functionality. The SupplyBoxProductsRepository was modified to return a pageable result.