Add KaspiToken update time and API URL adjustments
The changes included the addition of an update time field to KaspiToken. This timestamp helps in keeping track of the last update time. In addition, the Kaspi and File APIs have been parameterized in the application.yml file to make them configurable. The Seller API endpoint has been improved to use authentication token for user identification.