Update query conditions in several repositories
The commit includes changes in the 'KaspiOrderRepository', 'SupplyBoxProductsRepository', and 'ProductRepository'. The most notable changes are in the 'KaspiOrderRepository' where the logic of several queries was adjusted. In 'SupplyBoxProductsRepository' and 'ProductRepository', minor changes were made to better filter the results based on conditions. These adjustments aim to improve the accuracy of data retrieval.