Remove unnecessary comments and update phone number validation
Removed some redundant comments in the 'OrderServiceImpl.java' file and updated all instances of phone number validation in 'EmployeeCreateRequest.java' and 'EmployeeUpdateRequest.java' files. The updated validation checks for the length of phone numbers to be between 10 and 11 digits, this change allows more flexibility and meets the common phone number formatting standards.