Refactor product retrieval and mapping logic
The code has been refactored to optimize the retrieval and mapping of products. It now includes an additional retrieval of product price and supply box info based on product Ids. This attempt replaces the earlier approach that was focused on keycloak Id. Changes were also made in the ProductMapper and Product classes to correctly interpret and represent supply box info. Changes aim to improve product-related operations and ensure data retrieval consistency.