Add new features and modify existing ones for product handling
New classes for handling product categories and their respective repository services have been introduced. Changes have also been implemented to the handling of order and assembly states, and handling of product units and their types. This refactor also introduces changes in the DTO and response entities to align with the new handling in the services.