Add new repository methods and update service implementations
Two new methods 'findByArticle' and 'findByArticleAndOrderCode' are added to the 'SupplyBoxProductsRepository'. Meanwhile, the 'PackageServiceImpl' class is updated to use the new repository method, enhancing the specificity of the search function. Changes are also made to the 'SupplyServiceImpl' class to include validations for product and store, improving the robustness of the service.