Add create-supply.bpmn diagram
The diagram depicts the process of creating a new supply. It involves tasks such as generating reports, uploading files to GCS, and checking availability by box type and store time. The process also handles various error scenarios such as a city or store not existing and the user, city or store being disabled.