Update entities and services to use Set instead of List
This commit changes the usage from List to Set in several entities. It also updates the associated services to work with the new datatype. The purpose of these changes is to avoid potential duplication and improve performance. EqualsAndHashCode annotations have been added to properly handle the new datatype as well.