Refactor packaging process for individual products
The packaging logic has been revamped to support a start and finish process for individual products. The "PackageProductRequest" class has been removed, and the previous "packageProduct" method in the PackageService and PackageServiceImpl was replaced with "packageProductStart" and "packageProductFinish" methods. New corresponding API endpoints were added to the PackageController. This commit also introduces the "PACKING" state in the ProductStateInStore enumeration to differentiate between products that are ready for packaging and those being packed at the moment.